Entry(Day 25-26)
Worked:
1 hour-Or More(7:20ish-8:40ish)
         Forgot honestly how long I worked, the lab was fun.
Summary:
COMPLETED
I also added a list for each multiple of 3 and 5 that can be printed.
I did enjoy the lab entirely. 
Process:
I originally believed I would create a list that includes every number. Then I would take every value and compare it to 3 and 5, deciding if it is multiple.
I ended up with the idea that every number of times 3 is multiple.
Then with simple while and if statements, I figured it out.