Update Final Project (1.11.2021)
Time(1.10.2021)
6pm - 7:20pm

I was disappointed with the progress that I was making on problem 3. I figured out how to find all the prime numbers between 0 and a number. The next step would be configuring the loop to find the prime numbers whose product is the assigned value. I tried several times, but I do not fully understand modulos. Hopefully, the more I understand them, the easier problem with primes will be. 
Luckily, I have created an algorithm that makes a list of all the prime numbers within a limit. This hopefully will be applied to future projects and make the overall process easier.  I also continued the defined functions for each problem, I love this process, it makes progress seem easier.