Update Final Project(1.12.2020)
Overall the project was quite difficult. Figure a function in which I can not only produce a list of factors in between the numbers but also the prime composition of the factors. 
8:00pm-9-20pm
It did take longer than I wanted to figure out this project, But learning how to use modules is quite hard. It took extra time to manipulate modules to compete for prime factorization.