Update Final Project(1.21.2021)
This might be one of the hardest projects I have done. The progress that is shown is very minimal and it is frustrating. I am determined to finish problem 4. I know I am close. I have the filtering with the for loops and the if statements of digits that are larger than 6. How I need to confirm the numbers are the same that they are backward. Once I finish that I want to add it to a list. One major issue was I was not aware of the spilt function in python. Once I learned about it though I started making real progress.
Time: 7:25-9:30
I spent a lot of time on it. I will take a break to breathe and step back. I know the answer is in front of me, I simply cannot see it.
I still love how all of my work is on the same file though. It take me feel very professional.