Update Final Project
FINALLY FIGURED NUMBER 4!!!!!!!
I learned a lot. I did not know you could have multiple statements in the IF statement. Yes, I knew you could use two conjoined with a and statement. I was not aware you can chain them together. It took me a while to realize that. 
I was going crazy realizing that I could not finish this problem. I will not include how long I have worked on this session because it is sad. I enjoyed the work, it was a little frustrating.